action.py: generalize to multiple repositories
Generalize the script so it can check the history of any number of
projects in nrf/west.yml.

The main changes are:

- Capture the 'west ncs-loot --json' output for multiple projects, either as
  passed in via the command line or using a default list.

- Check histories for each project in a new <workspace>/oss-history
  directory. This makes the results easier to inspect, and has
  the side benefit that the local zephyr working tree is not rewritten
  anymore.

We're making use of 'west ncs-loot --json' output that was not
available at the time the script was originally written, specifically
the:

- path within the workspace for project we are checking
- nrf/west.yml commit we're checking
- zephyr/west.yml commit we're comparing against

This commit doesn't actually change the default set of projects yet,
but it is usable from the command line.

A subsequent commit will expand the default projects to check.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>