action.py: add --zephyr-merge-base option
This is needed when sdk-zephyr is based off of a zephyr branch that is
not 'main', like v2.7-branch. In that case, computing the 'git
merge-base' of the current sdk-zephyr manifest-rev against 'main' may
produce incorrect results depending on commit history.

To be able to solve such problems, add a --zephyr-merge-base option to
action.py that lets the user manually specify the merge base we should
be rebasing on top of.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>