action.py: add --no-user-config option
The script typically overrides the git user.name and user.email
configuration options in every git clone it creates while doing its
work, setting them to "oss-history" and "bot" respectively. This is
needed to avoid errors on environments where these variables are not
set, like those which can occur in CI containers.

When we are using this script to actually create the rebase tree,
though, we usually want to keep the default user.name and user.email
which correspond to the engineer doing the rebasing. This makes the
commiter information in the resulting 'git cherry-pick' commits
correct, so we aren't pushing nonsense into the public trees for
releases.

Add a --no-user-config option that overrides the default behavior to
make this happen.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>