action.py: add --quiet-subprocesses option
The default output from this script shows information about every
subprocess that runs, along with its stderr and often its stdout. This
is because as its primary purpose is to be run from a GitHub action,
we want all the necessary details to appear in the logs by default.

However, the script can also be used as part of the release process
by rebasing the OSS trees we've forked from upstream zephyr using
west. In this case, the defaults are too verbose and the script
output is overwhelming.

To make it easier to use the script in this way, add a
--quiet-subprocesses option that silences the extra information from
helper commands which run subprocesses.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>