action.py: merge strategy tweaks when cherry-picking
Use an explicit 'ort' merge strategy when cherry-picking, and make
sure that we provide the 'ours' merge strategy option to this
strategy. (We set the strategy explicitly to future-proof against
possible changes of the default strategy later to another one that
doesn't have the 'ours' option.)

See source code comments for a rationale of why we are adding the
'ours' strategy option. The short story is that this seems to help in
certain scenarios when the same change has been applied in both an OOT
and upstream commit, where the current script's behavior produces
false negatives (that is, it errors out when checking upmerges that
are in fact OK, and can safely be 'rebased' by adopting this new merge
strategy option).

I admit that this is a bit of a guess of how to fix a problem, but it
seems to help, and it "can't hurt".

This change "can't hurt" NCS since this script is still checking that
the resulting string of cherry-picks with the new option still produce
an empty diff from the current state of the tree. So as long as it's
successful when testing upmerges, we can still use it when doing an
automated history rewrite, with the desired results.

The worst thing that could happen is that this change starts to
produce errors of other kinds in future upmerges. I'm going to be
optimistic and hope that doesn't happen. However, if it does, we can
revisit and dig deeper into if there is a better strategy for
automated history maintenance in our open source forks.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>