action.py: forbid trailing '...' in shortlogs
We previously inherited short log line length limits in CI
from usptream projects. For example, upstream zephyr's .gitlint
restricted shortlogs to 75 characters, which caused errors
if the upstream shortlog plus a sauce tag like '[nrf fromtree]'
exceeded the limit.

The former workaround to this was to replace some trailing characters
in the upstream shortlog with '...' in NCS.

This is no longer needed since NCS forks of such repositories should
be extending any such limits to allow for our sauce tags, and the
current 'west ncs-loot' output for NCS at time of writing does not
show any commit shortlogs with a trailing '...'.

This makes it a good time to enforce a new "no ... at end of
shortlogs" rule.

The problem with the trailing ellipses is that they interfere with the
heuristics used by 'west ncs-compare' to match OOT commits with
upstream counterparts that are being brought in by an upmerge. Getting
rid of them therefore improves upmerge ergonomics.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>