action.py: s/shortlog/title/
I have incorrectly been referring to git commit titles as shortlogs.
Though git commit titles do appear in the 'git shortlog' output,
a git commit title is not a shortlog.

Fix it by replacing the terminology. There is one external area where
the JSON format in sdk-nrf may generate a 'shortlogs' string. We'll
fix that separately, but for now, add a shim to work with both.
Eventually we can drop the shim.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>